pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "news_crawler"
version = "0.1.0"
description = "Archive news articles (content + images) for offline viewing."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "UNLICENSED"}
authors = [{name = "news_crawler"}]
dependencies = [
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.0",
  "readability-lxml>=0.8.1",
  "lxml>=4.9.0",
]

[project.optional-dependencies]
render = ["playwright>=1.43.0"]

[project.scripts]
news_crawler = "news_crawler.cli:main"

[tool.setuptools]
packages = ["news_crawler"]

